provides syntax to use minimalkb models
allows the user to define local contexts to work with wanted models. the method active_models([ my_models ]) set my_models as active models to work with, and return object ModelContext to reset the default model (default) when the context is leaved.

the test.py code in testing/ provides unitest sequence function to test the code brought by this modification